# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Outlets Service",
    "description": "Manages press outlets (publications) and their domain authority data",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
    "schemas": {
      "CreateOutlet": {
        "type": "object",
        "properties": {
          "outletName": {
            "type": "string",
            "minLength": 1
          },
          "outletUrl": {
            "type": "string",
            "format": "uri"
          },
          "outletDomain": {
            "type": "string",
            "minLength": 1
          },
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "whyRelevant": {
            "type": "string"
          },
          "whyNotRelevant": {
            "type": "string"
          },
          "relevanceScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "overalRelevance": {
            "type": "string"
          },
          "relevanceRationale": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "ended",
              "denied"
            ],
            "default": "open"
          }
        },
        "required": [
          "outletName",
          "outletUrl",
          "outletDomain",
          "campaignId",
          "whyRelevant",
          "whyNotRelevant",
          "relevanceScore"
        ]
      },
      "UpdateOutlet": {
        "type": "object",
        "properties": {
          "outletName": {
            "type": "string",
            "minLength": 1
          },
          "outletUrl": {
            "type": "string",
            "format": "uri"
          },
          "outletDomain": {
            "type": "string",
            "minLength": 1
          },
          "whyRelevant": {
            "type": "string"
          },
          "whyNotRelevant": {
            "type": "string"
          },
          "relevanceScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "overalRelevance": {
            "type": "string"
          },
          "relevanceRationale": {
            "type": "string"
          }
        }
      },
      "UpdateOutletStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "ended",
              "denied"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "BulkCreateOutlets": {
        "type": "object",
        "properties": {
          "outlets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "outletName": {
                  "type": "string",
                  "minLength": 1
                },
                "outletUrl": {
                  "type": "string",
                  "format": "uri"
                },
                "outletDomain": {
                  "type": "string",
                  "minLength": 1
                },
                "campaignId": {
                  "type": "string",
                  "format": "uuid"
                },
                "whyRelevant": {
                  "type": "string"
                },
                "whyNotRelevant": {
                  "type": "string"
                },
                "relevanceScore": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 100
                },
                "overalRelevance": {
                  "type": "string"
                },
                "relevanceRationale": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "open",
                    "ended",
                    "denied"
                  ],
                  "default": "open"
                }
              },
              "required": [
                "outletName",
                "outletUrl",
                "outletDomain",
                "campaignId",
                "whyRelevant",
                "whyNotRelevant",
                "relevanceScore"
              ]
            }
          }
        },
        "required": [
          "outlets"
        ]
      },
      "SearchOutlets": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1
          },
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "limit": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "default": 20
          }
        },
        "required": [
          "query"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authenticating requests"
      }
    }
  },
  "paths": {
...
    "/outlets": {
      "post": {
        "summary": "Create outlet (upsert by outlet_url)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOutlet"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Outlet created"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List outlets with filters",
        "parameters": [
          {
            "in": "query",
            "name": "campaignId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "ended",
                "denied"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of outlets"
          }
        }
      }
    },
    "/outlets/{id}": {
      "get": {
        "summary": "Get outlet by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Outlet found"
          },
          "404": {
            "description": "Outlet not found"
          }
        }
      },
      "patch": {
        "summary": "Update outlet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOutlet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Outlet updated"
          },
          "404": {
            "description": "Outlet not found"
          }
        }
      }
    },
    "/outlets/{id}/status": {
      "patch": {
        "summary": "Update outlet status",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "campaignId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOutletStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/outlets/bulk": {
      "post": {
        "summary": "Bulk upsert outlets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateOutlets"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Outlets created"
          }
        }
      }
    },
    "/outlets/search": {
      "post": {
        "summary": "Search outlets by name/url",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchOutlets"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results"
          }
        }
      }
    },
...
  }
}